### 2023-08-30 #75 @sdwagner
[Re] When Does Label Smoothing Help?
**Original article:** Rafael Müller, Simon Kornblith, and Geoffrey E. Hinton. "When does label smoothing help?." Advances in neural information processing systems 32 (2019). (https://arxiv.org/pdf/1906.02629.pdf)

**Scientific domain:** Machine Learning
**Suggested editor:** Georgios Detorakis or Koustuv Sinha
