+++ artifact/dockerfile
@@ -1,0 +1,15 @@
+FROM python:3.9.13-buster
+#FROM pytorch/pytorch:2.4.1-cuda11.8-cudnn9-runtime
+
+
+COPY requirements.txt /docker_build/
+
+
+RUN pip install --upgrade pip && \
+    pip install -r /docker_build/requirements.txt
+
+
+
+WORKDIR /workspace
+# change dir to /workspace
+
