+++ datasets/datasets.py
@@ -16,15 +16,15 @@ def cifar10_training_dataset():
     training_data = torchvision.datasets.CIFAR10(root='./data', train=True,download=True, transform=training_transform)
     return training_data
 
-# Loading the CIFAR10 validation datasets and applying transformation
-def cifar10_validation_dataset():
+# Loading the CIFAR10 test datasets and applying transformation
+def cifar10_test_dataset():
     test_transform = transforms.Compose([
         transforms.CenterCrop(24), 
         transforms.ToTensor(), 
         transforms.Normalize((0.4923172, 0.48307145, 0.4474483),(0.24041407,0.23696952,0.25565723),True)
     ])
-    validation_data = torchvision.datasets.CIFAR10(root='./data', train=False,download=True, transform=test_transform)
-    return validation_data
+    test_data = torchvision.datasets.CIFAR10(root='./data', train=False,download=True, transform=test_transform)
+    return test_data
 
 # Loading the CIFAR10 training dataloader with a given batch_size
 def cifar10_training_loader(batch_size: int = 128):
@@ -32,11 +32,11 @@ def cifar10_training_loader(batch_size: int = 128):
     training_loader = t_data.DataLoader(training_data, batch_size=batch_size, shuffle=True, num_workers=2)
     return training_loader
 
-# Loading the CIFAR10 validation dataloader
-def cifar10_validation_loader():
-    validation_data = cifar10_validation_dataset()
-    validation_loader = t_data.DataLoader(validation_data, batch_size=len(validation_data), shuffle=True, num_workers=2)
-    return validation_loader
+# Loading the CIFAR10 test dataloader
+def cifar10_test_loader():
+    test_data = cifar10_test_dataset()
+    test_loader = t_data.DataLoader(test_data, batch_size=len(test_data), shuffle=True, num_workers=2)
+    return test_loader
 
 # Loading the CIFAR100 training datasets and applying transformation
 def cifar100_training_dataset():
@@ -50,15 +50,15 @@ def cifar100_training_dataset():
     training_data = t_data.Subset(training_data, range(40000))
     return training_data
 
-# Loading the CIFAR100 validation datasets and applying transformation
-def cifar100_validation_dataset():
+# Loading the CIFAR100 test datasets and applying transformation (using validation set according to Müller et al.)
+def cifar100_test_dataset():
     test_transform = transforms.Compose([
         transforms.ToTensor(), 
         transforms.Normalize([0.5071, 0.4865, 0.4409], [0.2673, 0.2564, 0.2761])
     ])
-    validation_data = torchvision.datasets.CIFAR100(root='./data', train=True,download=True, transform=test_transform)
-    validation_data = t_data.Subset(validation_data, range(40000, 50000))
-    return validation_data
+    test_data = torchvision.datasets.CIFAR100(root='./data', train=True,download=True, transform=test_transform)
+    test_data = t_data.Subset(test_data, range(40000, 50000))
+    return test_data
     
 # Loading the CIFAR100 training dataloader with a given batch_size
 def cifar100_training_loader(batch_size: int = 128):
@@ -66,11 +66,11 @@ def cifar100_training_loader(batch_size: int = 128):
     training_loader = t_data.DataLoader(training_data, batch_size=batch_size, shuffle=True, num_workers=2)
     return training_loader
 
-# Loading the CIFAR10 validation dataloader
-def cifar100_validation_loader():
-    validation_data = cifar100_validation_dataset()
-    validation_loader = t_data.DataLoader(validation_data, batch_size=len(validation_data), shuffle=True, num_workers=2)
-    return validation_loader
+# Loading the CIFAR100 test dataloader
+def cifar100_test_loader():
+    test_data = cifar100_test_dataset()
[…]
