+++ architectures/AlexNet.py
@@ -78,7 +78,7 @@ class AlexNet(nn.Module):
         return self.classifier[5], self.classifier[6]
     
     # Training Loop utilizing the standard training loop of trainer.py
-    def train_model(self, device, training_loader, validation_loader, label_smoothing_factor, epochs):
+    def train_model(self, device, training_loader, test_loader, label_smoothing_factor, epochs):
         
         #Definition of loss:
         cross_entropy = nn.CrossEntropyLoss(label_smoothing=label_smoothing_factor)
@@ -90,4 +90,4 @@ class AlexNet(nn.Module):
         #Definition of scheduler
         scheduler = MultiStepLR(opt, [415, 830], 0.1)
         
-        trainer.train_model(self, device, training_loader, validation_loader, epochs, train_loss, opt, scheduler)
+        trainer.train_model(self, device, training_loader, test_loader, epochs, train_loss, opt, scheduler)

[…]
