### 2024-10-11 #75 @tuelwer
@mo-arvan Thanks for creating a dockerfile! Feel free to open a PR to integrate it into our repository 😊
