### 2024-10-11 913fd43
Calculate standard error
