### 2024-10-11 46e0d2e
update dockerfile and readme
