+++ report/content.tex
@@ -359,7 +359,7 @@ To establish a baseline, the authors first trained teacher networks $(M1)$ and s
 \begin{figure}[ht]
 \centering
 \includegraphics[width=0.8\textwidth]{figures/fc_knowledge_distillation.pdf}
-\caption{Performance of distillation from fully-connected architecture ($1200$ neurons) to fully-connected architecture ($800$ neurons) on MNIST.}
+\caption{Performance of distillation from fully-connected architecture ($1200$ neurons) to fully-connected architecture ($800$ neurons) on MNIST. Shaded areas show the standard error.}
 \label{fig:fc_dist}
 \end{figure}
 
@@ -371,7 +371,7 @@ The authors also conducted the above-mentioned experiment on convolutional netwo
 \begin{figure}[ht]
 \centering
 \includegraphics[width=0.8\textwidth]{figures/knowledge_distillation.pdf}
-\caption{Performance of distillation from ResNet-56 to AlexNet on CIFAR-10.}
+\caption{Performance of distillation from ResNet-56 to AlexNet on CIFAR-10. Shaded areas show the standard error.}
 \label{fig:conv_dist}
 
 \end{figure}
