### 2024-10-14 #75 @tuelwer
@mo-arvan Thanks again for your detailed comments! In the following we want to address each of the points that you raised:

1. **Confusion validation and test data:** We carefully double-checked our datasets and can confirm that all experiments were performed on the test split of each dataset:
	- For the datasets implemented by PyTorch, we set `train=False`, which corresponds to the test split (please refer to, e.g., [here](https://pytorch.org/vision/main/generated/torchvision.datasets.EMNIST.html#torchvision.datasets.EMNIST)).
	- For the CUB-200-2011 data we use the test split of the dataset which is defined in the file `train_test_split.txt`. The CUB-200-2011 dataset does not have a validation set.
	- For the Tiny ImageNet we use the split that is defined as validation split. We assume that the authors of the paper did this as well, since the test data split of the Tiny ImageNet dataset is *not labeled*.
We apologize for the confusion, and we have refactored the code accordingly.

2. **Uncertainty quantification:** We added confidence intervals for Figure 6 and 7.

3. **Color palette:** We have chosen the colors that were used in the original work to allow easy comparison of the experimental results.

4. **Edits:** We have incorporated the proposed changes into [our report](https://github.com/sdwagner/re-labelsmoothing/blob/main/report/article.pdf).
