### 2024-11-01 #75 @mo-arvan
Yes, the response addresses my main concerns. I was wrong about the validation/test splits.
