### 2025-02-17 #75 @rougier
We're very very late on this submission and based on the review by @mo-arvan, I propose the paper to be accepted. Woudl that be ok @gdetor
