### 2025-02-19 #75 @tuelwer
@rougier The source of the article can be found here: https://github.com/sdwagner/re-labelsmoothing/tree/main/report
